+++ test/js/revoke.js
@@ -13,6 +13,7 @@ var util = require('./acme-util');
 var forge = require('node-forge');
 var fs = require('fs');
 var request = require('request');
+var Acme = require('./acme');
 
 function main() {
   if (process.argv.length != 5) {
@@ -20,43 +21,23 @@ function main() {
     process.exit(1);
   }
   var certFile = process.argv[2];
-  console.log('Attempting to revoke', certFile);
   var key = crypto.importPemPrivateKey(fs.readFileSync(process.argv[3]));
+  var acme = new Acme(key);
   var certDER = fs.readFileSync(certFile)
+  if (certDER.toString().match(/-----BEGIN/)) {
+    console.log('Got PEM, expected DER:', certFile);
+    process.exit(1);
+  }
   var revokeUrl = process.argv[4];
   var certDERB64URL = util.b64enc(new Buffer(certDER))
-  var revokeMessage = JSON.stringify({
-    resource: "revoke-cert",
+  console.log('Attempting to revoke', certFile);
+  acme.post(revokeUrl, {
+    resource: 'revoke-cert',
     certificate: certDERB64URL
-  });
-  console.log('Requesting revocation:', revokeMessage)
-
-  request.head(revokeUrl, function(error, response, body) {
-    if (error) {
-      console.log(error);
-      process.exit(1);
-    } else if (response.statusCode != 200) {
-      console.log("Got non-200 response asking for nonce (non-fatal):", response.statusCode);
+  }, function(err, response, body) {
+    if (!err && response.statusCode === 200) {
+      console.log('Success');
     }
-    var nonce = response.headers["replay-nonce"];
-    if (!nonce) {
-      console.log("Server HEAD response did not include a replay nonce");
-      process.exit(1);
-    }
-
-    var jws = crypto.generateSignature(key, new Buffer(revokeMessage), nonce);
-    var payload = JSON.stringify(jws);
-
-    var req = request.post(revokeUrl, function(error, response) {
-      if (error || Math.floor(response.statusCode / 100) != 2) {
-        console.log('Error: ', error);
-        process.exit(1);
-      } else {
-        console.log("Revocation request successful.");
-      }
-    });
-    req.write(payload);
-    req.end();
   });
 }
 main();
